Thu, Sep 14, 2023, 3:18 PM +02:00

+++ filter-less/helpers.c
@@ -102,9 +102,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             temp[i][j] = image[i][j];
         }
     }
-    float totalRed, totalGreen, totalBlue;
+    int totalRed, totalGreen, totalBlue;
     totalRed = totalGreen = totalBlue = 0;
-    float counter = 0;
+    int counter = 0;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
